Try except sur les installations

+++ mac/installation.py
@@ -21,32 +21,36 @@ def executeAsUser(command):
     os.system("sudo -u " + currentUser + " " + command)
 
 def macupdate():
-    print("Installation des mises à jour de MacOS en général Xcode inclus")
-    os.system("softwareupdate --install -a")
+    try:
+        print("\n\n\n####################################################  Installation des mises à jour de MacOS en général Xcode inclus")
+        os.system("softwareupdate --install -a")
+    except:
+        print("Erreur lors de l'installation des mises à jour de MacOS")
 
 def flutter():
-    currentUser = os.popen("id -un").read().strip()
-    print("Installation de flutter as " + currentUser)
-    if os.system("which flutter") == 0:
-        print("Flutter est déjà installé")
-        # sudo -u "$currentUser" defaults write com.apple.dock orientation left
-        executeAsUser("flutter upgrade")
-    else:
-        print("Installation de flutter nouveau")
-        telecharge(
-            flutterURL,
-            "/Users/"+userName+"/Downloads/flutter.zip")
-        os.system("mkdir /opt/flutter")
-        os.system("chown -R " + userName + " /opt/flutter")
-        os.system("chmod 777 -R  /opt/flutter")
-        executeAsUser("unzip -qq /Users/"+userName+"/Downloads/flutter.zip -d /opt/")
-        os.system("chown -R " + userName + " /opt/flutter")
-        os.system("chgrp -R admin /opt/flutter")
-        os.system("chmod 777 -R  /opt/flutter")
-        # deplacer flutter dans /opt
-        # os.system("sudo mv flutter /opt/")
-        add_to_system_path("/opt/flutter/bin")
-        # ajouter flutter au path
+    try:
+        print("\n\n\n####################################################  Installation de flutter as " + userName)
+        if os.system("which flutter") == 0:
+            print("Flutter est déjà installé")
+            executeAsUser("flutter upgrade")
+        else:
+            print("Installation de flutter nouveau")
+            telecharge(
+                flutterURL,
+                "/Users/"+userName+"/Downloads/flutter.zip")
+            os.system("mkdir /opt/flutter")
+            os.system("chown -R " + userName + " /opt/flutter")
+            os.system("chmod 777 -R  /opt/flutter")
+            executeAsUser("unzip -qq /Users/"+userName+"/Downloads/flutter.zip -d /opt/")
+            os.system("chown -R " + userName + " /opt/flutter")
+            os.system("chgrp -R admin /opt/flutter")
+            os.system("chmod 777 -R  /opt/flutter")
+            # deplacer flutter dans /opt
+            # os.system("sudo mv flutter /opt/")
+            add_to_system_path("/opt/flutter/bin")
+            # ajouter flutter au path
+    except:
+        print("Erreur lors de l'installation de Flutter")
     #print("Etat de l'installation de Flutter")
     #os.system("flutter doctor")
 
@@ -57,37 +61,51 @@ def add_to_system_path(path):
             if path in line:
                 break
         else:  # not found, we are at the eof
-            file.write(path)  # append missing data
+            with open("/etc/paths", 'r') as original:
+                data = original.read()
+            with open("/etc/paths", 'w') as modified:
+                modified.write(path+"\n" + data)
+            #file.write(path)  # append missing data
 
 
 def rosetta():
-    print("Installation de Rosetta")
-    os.system("sudo softwareupdate --install-rosetta --agree-to-license")
+    try:
+        print("\n\n\n####################################################  Installation de Rosetta")
+        os.system("sudo softwareupdate --install-rosetta --agree-to-license")
+    except:
+        print("Erreur lors de l'installation de Rosetta")
 def homebrew():
-    print("Installation de Brew ou mise à jour")
+    print("\n\n\n####################################################  Installation de Brew ou mise à jour")
     # test if brew is installed
     if executeAsUser("which brew") == 0:
-        print("Brew est déjà installé")
+        print("\n\n\n####################################################  Brew est déjà installé")
         executeAsUser("brew update")
     else:
-        print("Installation de Brew")
+        print("\n\n\n####################################################  Installation de Brew")
         executeAsUser("/bin/bash -c \"$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)\"")
         add_to_system_path("/opt/homebrew/bin")
         executeAsUser("brew update")
     # install homebrew
 
+def install_ruby3():
+    print("\n\n\n####################################################  Installation de ruby 3")
+    os.system("brew install ruby")
+    add_to_system_path("/opt/homebrew/opt/ruby/bin")
+    os.system("gem update --system")
+
 def cocoapods():
-    print("Installation de cocoapods / mise à jour")
-    os.system("sudo gem update")
+    print("\n\n\n####################################################  Installation de cocoapods / mise à jour")
+    os.system("gem install cocoapods")
+    os.system("gem update")
     print("Mise à jour de cocoapods")
-    os.system("sudo gem install cocoapods")
     # cocoapods cannot be run as root
     executeAsUser("pod repo update")
 
 def xcode():
-    print("Configure Xcode")
+    print("\n\n\n####################################################  Configure Xcode")
     # configure Xcode
-    os.system("sudo xcode-select --switch /Applications/Xcode.app/Contents/Developer")
+    os.system("sudo xcode-select –-install")
+    # os.system("sudo xcode-select --switch /Applications/Xcode.app/Contents/Developer")
     os.system("sudo xcodebuild -runFirstLaunch")
     os.system("open -a XCode")
 
@@ -97,24 +115,29 @@ def telecharge(url, destination):
 
 def installDansApplication(url, tempFile, volumeName, applicationName):
     # test first if the application is already installed
+    if os.system("ls /Applications | grep \"" + applicationName + "\"") == 0:
+        print(applicationName + " est déjà installé, enlever si vous voulez réinstaller")
+        return
+    try:
+        os.system("rm -rf \"/Applications/" + applicationName + "\"")
+        print("Telechargement de " + applicationName)
+        telecharge(url, tempFile)
 
-    os.system("rm -rf \"/Applications/" + applicationName + "\"")
-    print("Telechargement de " + applicationName)
-    telecharge(url, tempFile)
+        mountCommand = "hdiutil attach -mountpoint \"/Volumes/" + volumeName + "\" " + tempFile + ("  -quiet")
+        print("Montage du volume de " + mountCommand)
+        os.system(mountCommand)
 
-    mountCommand = "hdiutil attach -mountpoint \"/Volumes/" + volumeName + "\" " + tempFile + ("  -quiet")
-    print("Montage du volume de " + mountCommand)
-    os.system(mountCommand)
+        copyCommand = "cp -Rf \"/Volumes/" + volumeName + "/" + applicationName + "\"" " /Applications"
+        print("Copie de l'application " + copyCommand)
+        os.system(copyCommand)
 
-    copyCommand = "cp -Rf \"/Volumes/" + volumeName + "/" + applicationName + "\"" " /Applications"
-    print("Copie de l'application " + copyCommand)
-    os.system(copyCommand)
-
-    #print("Demontage de l'image \"" + applicationName + "\" dans Applications")
-    os.system("hdiutil detach \"/Volumes/" + volumeName+ "\" -quiet ")
-    os.system("rm " + tempFile)
-    #print("Demarrage de l'application " + applicationName )
-    os.system("open -a \"" + applicationName+ "\" ")
+        #print("Demontage de l'image \"" + applicationName + "\" dans Applications")
+        os.system("hdiutil detach \"/Volumes/" + volumeName+ "\" -quiet ")
+        os.system("rm " + tempFile)
+        #print("Demarrage de l'application " + applicationName )
+        os.system("open -a \"" + applicationName+ "\" ")
+    except:
+        print("Erreur lors de l'installation de " + applicationName)
 
 flutterLocation = "/opt/flutter"
 flutterURL = "https://storage.googleapis.com/flutter_infra_release/releases/stable/macos/flutter_macos_arm64_3.16.1-stable.zip"
@@ -155,10 +178,15 @@ installDansApplication(
     "GitKraken",
     "GitKraken.app")
 
+
+add_to_system_path("pipo")
 #githubDesktop() TODO fix it is broken
 macupdate()
 xcode()
 homebrew()
+install_ruby3()
+os.system("ruby -v")
+os.system("gem -v")
 cocoapods()
 rosetta()
 flutter()
@@ -166,5 +194,6 @@ flutter()
 os.system("flutter --version")
 os.system("pod --version")
 os.system("/usr/bin/xcodebuild -version")
+os.system("ruby -v")
